+++ hack/capi.go
@@ -1,8 +1,0 @@
-// +build capi
-
-package hack
-
-import (
-	_ "sigs.k8s.io/cluster-api"
-	_ "sigs.k8s.io/cluster-api-provider-azure"
-)