+++ pkg/apis/core/v1alpha1/cert_types.go
@@ -1,21 +1,15 @@
 package v1alpha1
 
 import (
-	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
-	"github.com/giantswarm/apiextensions/v3/pkg/crd"
 )
 
 const (
 	kindCertConfig              = "CertConfig"
 	certConfigDocumentationLink = "https://docs.giantswarm.io/reference/cp-k8s-api/certconfigs.core.giantswarm.io/"
 )
-
-func NewCertConfigCRD() *v1.CustomResourceDefinition {
-	return crd.LoadV1(group, kindCertConfig)
-}
 
 // NewCertConfigTypeMeta returns the type part for the metadata section of a
 // CertConfig custom resource.