+++ pkg/apis/infrastructure/v1alpha2/aws_cluster_types.go
@@ -1,21 +1,15 @@
 package v1alpha2
 
 import (
-	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
-	"github.com/giantswarm/apiextensions/v3/pkg/crd"
 )
 
 const (
 	kindAWSCluster              = "AWSCluster"
 	awsClusterDocumentationLink = "https://docs.giantswarm.io/reference/cp-k8s-api/awsclusters.infrastructure.giantswarm.io/"
 )
-
-func NewAWSClusterCRD() *v1.CustomResourceDefinition {
-	return crd.LoadV1(group, kindAWSCluster)
-}
 
 func NewAWSClusterTypeMeta() metav1.TypeMeta {
 	return metav1.TypeMeta{