+++ pkg/apis/infrastructure/v1alpha2/aws_control_plane_types.go
@@ -1,21 +1,15 @@
 package v1alpha2
 
 import (
-	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
-	"github.com/giantswarm/apiextensions/v3/pkg/crd"
 )
 
 const (
 	kindAWSControlPlane              = "AWSControlPlane"
 	awsControlPlaneDocumentationLink = "https://docs.giantswarm.io/reference/cp-k8s-api/awscontrolplanes.infrastructure.giantswarm.io/"
 )
-
-func NewAWSControlPlaneCRD() *v1.CustomResourceDefinition {
-	return crd.LoadV1(group, kindAWSControlPlane)
-}
 
 func NewAWSControlPlaneTypeMeta() metav1.TypeMeta {
 	return metav1.TypeMeta{