+++ pkg/apis/infrastructure/v1alpha2/g8s_control_plane_types.go
@@ -2,21 +2,15 @@ package v1alpha2
 
 import (
 	corev1 "k8s.io/api/core/v1"
-	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
-	"github.com/giantswarm/apiextensions/v3/pkg/crd"
 )
 
 const (
 	kindG8sControlPlane              = "G8sControlPlane"
 	g8sControlPlaneDocumentationLink = "https://docs.giantswarm.io/reference/cp-k8s-api/g8scontrolplanes.infrastructure.giantswarm.io/"
 )
-
-func NewG8sControlPlaneCRD() *v1.CustomResourceDefinition {
-	return crd.LoadV1(group, kindG8sControlPlane)
-}
 
 func NewG8sControlPlaneTypeMeta() metav1.TypeMeta {
 	return metav1.TypeMeta{