+++ pkg/apis/release/v1alpha1/release_types.go
@@ -1,11 +1,9 @@
 package v1alpha1
 
 import (
-	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
-	"github.com/giantswarm/apiextensions/v3/pkg/crd"
 )
 
 const (
@@ -24,10 +22,6 @@ var (
 
 func (r ReleaseState) String() string {
 	return string(r)
-}
-
-func NewReleaseCRD() *v1.CustomResourceDefinition {
-	return crd.LoadV1(group, kindRelease)
 }
 
 func NewReleaseTypeMeta() metav1.TypeMeta {