+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added the `ui.giantswarm.io/original-organization-name` annotation
+
+- Add CAPI CRDs.
+- Added the `ui.giantswarm.io/original-organization-name` annotation.
 
 ### Changed
 - Updated the documentation for the `alpha.aws.giantswarm.io/aws-subnet-size` annotation to explain the current behaviour